refactored the code. fixed bugs
