+++ hrms-client/src/App.jsx
@@ -1,11 +1,14 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
+
+import "./assets/css/common.css";
+
 import Home from "./components/Home";
 import Navbar from "./components/Navbar";
-import "./assets/css/common.css";
 import About from "./components/About";
 import Login from "./components/Login";
 import Footer from "./components/Footer";
+import Register from "./components/Register";
 
 const App = () => {
     return (
@@ -16,6 +19,7 @@ const App = () => {
                     <Route path='/' element={<Home />} />
                     <Route path='/about' element={<About />} />
                     <Route path='/login' element={<Login />} />
+                    <Route path='/register' element={<Register />} />
                 </Routes>
             </div>
             <Footer />
