+++ hrms-client/src/components/Register.jsx
@@ -1,0 +1,6 @@
+import "../assets/css/register.css";
+const Register = () => {
+    return <div>this is register component</div>;
+};
+
+export default Register;
